refactor: create widgets folder src/widgets and move sidebar, modals,layout, header to it
- Update imports for header, layout, modals, sidebar
