+++ src/features/index.ts
@@ -1,5 +1,3 @@
 export * from './auth'
-export * from './layout'
-export * from './modals'
 export * from './post'
 export * from './profile'
