+++ src/shared/ui/index.ts
@@ -1,8 +1,6 @@
 export * from './FormDatePicker'
 export * from './datePicker'
-export * from './header'
 export * from './languageSelect'
 export * from './profileStat'
 export * from './reCaptcha'
-export * from './sidebar'
 export * from './trans'
