+++ src/widgets/layout/layout.module.scss
