+++ src/widgets/modals/ui/confirmationModal/confirmationModal.module.scss
