+++ src/widgets/modals/ui/confirmationModal/index.ts
