+++ src/widgets/modals/ui/modalAddPhoto/index.ts
