+++ src/widgets/modals/ui/modalAddPhoto/modalAddPhoto.module.scss
