+++ src/widgets/sidebar/sidebar.module.scss
